updated docs
Updated the documentation. Also moved around and renamed some internal functions while doing it.
