+++ docs/modules.rst
@@ -12,7 +12,6 @@ module can be used, but rather how it works internally.
     modules/_mfv2d
     modules/boundary
     modules/continuity
-    modules/element
     modules/eval
     modules/kform
     modules/mimetic2d
