+++ python/mfv2d/mimetic2d.py
@@ -35,7 +35,6 @@ from mfv2d._mfv2d import (
     Manifold2D,
     Mesh,
     Surface,
-    compute_gll,
     dlagrange1d,
     lagrange1d,
 )
@@ -451,145 +450,6 @@ def vtk_lagrange_ordering(order: int) -> npt.NDArray[np.uint32]:
     )
 
 
-@dataclass(eq=False)
-class Element2D:
-    """General 2D element."""
-
-    parent: ElementNode2D | None
-
-
-@dataclass(eq=False)
-class ElementNode2D(Element2D):
-    """Two dimensional element that contains children."""
-
-    child_bl: Element2D
-    child_br: Element2D
-    child_tl: Element2D
-    child_tr: Element2D
-
-    def children(self) -> tuple[Element2D, Element2D, Element2D, Element2D]:
-        """Return children of the element ordered."""
-        return (self.child_bl, self.child_br, self.child_tr, self.child_tl)
-
-
-@dataclass(eq=False)
-class ElementLeaf2D(Element2D):
-    """Two dimensional square element.
-
-    This type facilitates operations related to calculations which need
-    to be carried out on the reference element itself, such as calculation
-    of the mass and incidence matrices, as well as the reconstruction of
-    the solution.
-
-    Parameters
-    ----------
-    order_h : int
-        Order of the basis functions used for the nodal basis in the first dimension.
-    order_v : int
-        Order of the basis functions used for the nodal basis in the second dimension.
-    bottom_left : (float, float)
-        Coordinates of the bottom left corner.
-    bottom_right : (float, float)
-        Coordinates of the bottom right corner.
-    top_right : (float, float)
-        Coordinates of the top right corner.
-    top_left : (float, float)
-        Coordinates of the top left corner.
-    """
-
-    order: int
-
-    bottom_left: tuple[float, float]
-    bottom_right: tuple[float, float]
-    top_right: tuple[float, float]
-    top_left: tuple[float, float]
-
-    def divide(
-        self,
-        order_bl: int,
-        order_br: int,
-        order_tl: int,
-        order_tr: int,
-    ) -> tuple[
-        ElementNode2D,
-        tuple[tuple[ElementLeaf2D, ElementLeaf2D], tuple[ElementLeaf2D, ElementLeaf2D]],
-    ]:
-        """Divide the element into four child elements of the specified order.
-
-        Parameters
-        ----------
-        order_bl : int
-            Order of the bottom left element.
-        order_br : int
-            Order of the bottom right element.
-        order_tl : int
-            Order of the top left element.
-        order_tr : int
-            Order of the top right element.
-
-        Returns
-        -------
-        ElementNode2D
-            Parent element which contains the nodes.
-
-        (2, 2) tuple of ElementLeaf2D
-            Child elements of the same order as the element itself. Indexing the
-            tuple will give bottom/top for the first axis and left/right for the
-            second.
-        """
-        bottom_mid = (np.array(self.bottom_left) + np.array(self.bottom_right)) / 2
-        left_mid = (np.array(self.bottom_left) + np.array(self.top_left)) / 2
-        right_mid = (np.array(self.bottom_right) + np.array(self.top_right)) / 2
-        top_mid = (np.array(self.top_left) + np.array(self.top_right)) / 2
-        center_mid = (
-            np.array(self.bottom_left)
-            + np.array(self.bottom_right)
-            + np.array(self.top_left)
-            + np.array(self.top_right)
-        ) / 4
-        btm_l = ElementLeaf2D(
-            None,
-            order_bl,
-            self.bottom_left,
-            tuple(bottom_mid),
-            tuple(center_mid),
-            tuple(left_mid),
-        )
-        btm_r = ElementLeaf2D(
-            None,
-            order_br,
-            tuple(bottom_mid),
-            self.bottom_right,
-            tuple(right_mid),
-            tuple(center_mid),
-        )
-        top_r = ElementLeaf2D(
-            None,
-            order_tr,
-            tuple(center_mid),
-            tuple(right_mid),
-            self.top_right,
-            tuple(top_mid),
-        )
-        top_l = ElementLeaf2D(
-            None,
-            order_tl,
-            tuple(left_mid),
-            tuple(center_mid),
-            tuple(top_mid),
-            self.top_left,
-        )
-
-        parent = ElementNode2D(self.parent, btm_l, btm_r, top_l, top_r)
-
-        btm_l.parent = parent
-        btm_r.parent = parent
-        top_l.parent = parent
-        top_r.parent = parent
-
-        return parent, ((btm_l, btm_r), (top_l, top_r))
-
-
 class FemCache:
     """Cache for integration rules and basis functions.
 
@@ -861,26 +721,6 @@ def element_node_children_on_side(
     return int(children[i_begin]), int(children[i_end])
 
 
-def get_element_side_nodes(
-    mesh: Mesh, element: int, side: ElementSide
-) -> npt.NDArray[np.float64]:
-    """Get nodes on the specified side."""
-    children = mesh.get_element_children(element)
-    if children is None:
-        orders = mesh.get_leaf_orders(element)
-        order = orders[(side.value - 1) & 1]
-        return compute_gll(order)[0]
-
-    c1, c2 = element_node_children_on_side(side, children)
-
-    return np.concatenate(
-        (
-            (get_element_side_nodes(mesh, c1, side) - 1) / 2,
-            (get_element_side_nodes(mesh, c2, side) + 1) / 2,
-        )
-    )
-
-
 def element_boundary_dofs(
     side: ElementSide, order: UnknownFormOrder, order_1: int, order_2: int
 ) -> npt.NDArray[np.uint32]:
@@ -993,77 +833,6 @@ class ElementConstraint:
     coeffs: npt.NDArray[np.float64]
 
 
-def get_side_dof_nodes(
-    mesh: Mesh,
-    element: int,
-    side: ElementSide,
-    order: UnknownFormOrder,
-    /,
-) -> list[ElementConstraint]:
-    """Get equations for obtaining DoFs on the side for the element.
-
-    Parameters
-    ----------
-    element_collection : ElementCollection
-        Collection of elements the ``element`` is in.
-
-    element : int
-        Element which to get DoF equations for.
-
-    side : ElementSide
-        Side of the element to get the DoFs for.
-
-    order : UnknownFormOrder
-        Order of unknown forms for which to get the form orders from. It
-        can only be ``UnknownFormOrder.FORM_ORDER_0`` or
-        ``UnknownFormOrder.FORM_ORDER_1``, since 2-forms do not have any
-        boundary DoFs.
-
-    Returns
-    -------
-    constraints : list of ElementConstraint
-        Specifiecation of which *leaf* elements are involved. These
-        are also ordered. The ``coeff`` member specifies positions
-        of these in the element.
-    """
-    children = mesh.get_element_children(element)
-    if children is not None:
-        c1: int
-        c2: int
-        c1, c2 = element_node_children_on_side(side, children)
-
-        dofs1 = get_side_dof_nodes(mesh, c1, side, order)
-        dofs2 = get_side_dof_nodes(mesh, c2, side, order)
-
-        if order == UnknownFormOrder.FORM_ORDER_0:
-            # Do not include the last row (DoF shared between the two)
-            # since it overlaps with M1. Otherwise it overconstrains.
-            dofs2[0] = ElementConstraint(
-                dofs2[0].i_e, dofs2[0].dofs[1:], dofs2[0].coeffs[1:]
-            )
-        elif order == UnknownFormOrder.FORM_ORDER_1:
-            # Still have to remove coeffs (since that stands for child nodes!)
-            dofs2[0] = ElementConstraint(dofs2[0].i_e, dofs2[0].dofs, dofs2[0].coeffs[1:])
-        else:
-            assert False
-
-        combined_dofs = [
-            ElementConstraint(dof.i_e, dof.dofs, (dof.coeffs - 1) / 2) for dof in dofs1
-        ] + [ElementConstraint(dof.i_e, dof.dofs, (dof.coeffs + 1) / 2) for dof in dofs2]
-
-        return combined_dofs
-
-    # This is a leaf
-    n1, n2 = mesh.get_leaf_orders(element)
-
-    indices = element_boundary_dofs(side, order, n1, n2)
-
-    side_orders = mesh.get_leaf_orders(element)
-    side_order = side_orders[(side.value - 1) & 1]
-
-    return [ElementConstraint(element, indices, compute_gll(side_order)[0])]
-
-
 def get_side_order(mesh: Mesh, element_idx: int, side: ElementSide, /) -> int:
     """Get order for the specified element boundary side.
 
@@ -1112,178 +881,6 @@ class Constraint:
     def __init__(self, rhs: float, *element_constraints: ElementConstraint) -> None:
         object.__setattr__(self, "rhs", float(rhs))
         object.__setattr__(self, "element_constraints", element_constraints)
-
-
-def get_side_dofs(
-    mesh: Mesh,
-    element: int,
-    side: ElementSide,
-    form_order: UnknownFormOrder,
-    output_order: int | None = None,
-    /,
-) -> tuple[Constraint, ...]:
-    """Get DoFs on the boundary in terms of leaf element DoFs.
-
-    Parameters
-    ----------
-    element_collection : ElementCollection
-        Collection of elements the ``element`` is in.
-
-    element : int
-        Element which to get DoF equations for.
-
-    side : ElementSide
-        Side of the element to get the DoFs for.
-
-    order : UnknownFormOrder
-        Order of unknown forms for which to get the form orders from. It
-        can only be ``UnknownFormOrder.FORM_ORDER_0`` or
-        ``UnknownFormOrder.FORM_ORDER_1``, since 2-forms do not have any
-        boundary DoFs.
-
-    Returns
-    -------
-    tuple of Constraint
-        Tuple of constraints, each of which specifies how the "normal" DoFs
-        on the element's boundary may be constructed from DoFs of the element.
-    """
-    self_order = get_side_order(mesh, element, side)
-
-    # If output order is not specified, use own order
-    if output_order is None:
-        output_order = self_order
-
-    if mesh.get_element_children(element) is None and output_order == self_order:
-        # fast track for leaf elements with no projection, since it should be identity
-        indices = element_boundary_dofs(side, form_order, *mesh.get_leaf_orders(element))
-        out_c = tuple(
-            Constraint(
-                0.0,
-                ElementConstraint(
-                    element, np.array([idx], np.uint32), np.ones(1, np.float64)
-                ),
-            )
-            for idx in indices
-        )
-        return out_c
-
-    dofs = get_side_dof_nodes(mesh, element, side, form_order)
-
-    self_nodes = compute_gll(self_order)[0]
-    input_nodes = np.concatenate([dof.coeffs for dof in dofs])
-
-    # Values of output basis (axis 1) at input points (axis 0)
-    # nodal_basis_vals and edge_basis_vals are maps from parent dofs to child dofs
-    nodal_basis_vals = lagrange1d(self_nodes, input_nodes)
-    if form_order == UnknownFormOrder.FORM_ORDER_0:
-        m = np.linalg.inv(nodal_basis_vals)
-
-    elif form_order == UnknownFormOrder.FORM_ORDER_1:
-        diffs = nodal_basis_vals[:-1, :] - nodal_basis_vals[+1:, :]
-        edge_basis_vals = np.stack(
-            [x for x in accumulate(diffs[..., i] for i in range(diffs.shape[-1] - 1))],
-            axis=-1,
-            dtype=np.float64,
-        )
-        m = np.linalg.inv(edge_basis_vals)
-
-    elif form_order == UnknownFormOrder.FORM_ORDER_2:
-        raise ValueError("2-forms have no boundary DoFs.")
-
-    else:
-        raise ValueError(f"Invalid for order {form_order=}.")
-
-    if self_order != output_order:
-        # Values of output basis (axis 1) at input points (axis 0)
-        # nodal_basis_vals and edge_basis_vals are maps from parent dofs to child dofs
-        output_nodes = compute_gll(output_order)[0]
-        map_nodal_basis_vals = lagrange1d(self_nodes, output_nodes)
-        if form_order == UnknownFormOrder.FORM_ORDER_0:
-            m = map_nodal_basis_vals @ m
-
-        elif form_order == UnknownFormOrder.FORM_ORDER_1:
-            diffs = map_nodal_basis_vals[:-1, :] - map_nodal_basis_vals[+1:, :]
-            map_edge_basis_vals = np.stack(
-                [
-                    x
-                    for x in accumulate(diffs[..., i] for i in range(diffs.shape[-1] - 1))
-                ],
-                axis=-1,
-                dtype=np.float64,
-            )
-            m = map_edge_basis_vals @ m
-
-        elif form_order == UnknownFormOrder.FORM_ORDER_2:
-            raise ValueError("2-forms have no boundary DoFs.")
-
-        else:
-            raise ValueError(f"Invalid for order {form_order=}.")
-
-    constraints: list[Constraint] = list()
-
-    vrow: npt.NDArray[np.float64]
-    for vrow in m:
-        col_offset = 0
-        elem_constraints: list[ElementConstraint] = list()
-
-        for elem_dofs in dofs:
-            cnt = elem_dofs.dofs.size
-            element_constraint = ElementConstraint(
-                elem_dofs.i_e, elem_dofs.dofs, vrow[col_offset : col_offset + cnt]
-            )
-            col_offset += cnt
-            elem_constraints.append(element_constraint)
-
-        assert col_offset == vrow.size
-        constraint = Constraint(0.0, *elem_constraints)
-        constraints.append(constraint)
-
-    return tuple(constraints)
-
-
-def get_corner_dof(mesh: Mesh, element: int, side: ElementSide, /) -> tuple[int, int]:
-    """Get element index and degree of freedom index for the corner of the element.
-
-    Parameters
-    ----------
-    element_collection : ElementCollection
-        Collection of elements the ``element`` is in.
-
-    element : int
-        Element which to get DoF equations for.
-
-    side : ElementSide
-        Side which begins with the corner that should be obtained.
-
-    Returns
-    -------
-    element_id : int
-        Index of the (leaf) element to which the corner belongs to.
-
-    dof_index : int
-        Index of the (0-form) degree of freedom which is in that corner.
-    """
-    children = mesh.get_element_children(element)
-    if children is None:
-        # Actual leaf
-        order_1, order_2 = mesh.get_leaf_orders(element)
-
-        if side == ElementSide.SIDE_BOTTOM:
-            idx = 0
-        elif side == ElementSide.SIDE_RIGHT:
-            idx = order_1
-        elif side == ElementSide.SIDE_TOP:
-            idx = (order_1 + 1) * order_2 + order_1
-        elif side == ElementSide.SIDE_LEFT:
-            idx = order_2 * (order_1 + 1)
-        else:
-            raise ValueError(f"Invalid side given by {side=}")
-
-        return (element, idx)
-
-    child = children[side.value - 1]
-
-    return get_corner_dof(mesh, child, side)
 
 
 def compute_leaf_dof_counts(
